Add solution of Q1068
Add solution of CodeUp's Question No.1068
This question can be solved with different way.
The first one is If statement.
'If' statement is easy to use but the code is getting longer and the lines are getting longer as well.
The easy way to solve is 'Switch' statement.
One thing you have to be careful is you can't designate range.