Add solution of Q1079, Q1080
Add solution of CodeUp's Question No.1079, No.1080
This question can be solved with for statement.
But if you use  while statement it'll be easy to solve.
If you know how much cycle times needed, for statement is more suit.
But if you dont'know how much cycle times needed, while statement is suit.